requirements.txt
pynacl
pycryptodome
x25519
fastcrc
